+++ crates/ryo3-serde/src/ser/mod.rs
@@ -5,7 +5,6 @@ pub(crate) mod py_serialize;
 #[expect(clippy::inline_always)]
 pub(crate) mod py_types;
 mod ry_types;
-mod traits;
 
 pub(crate) use context::PySerializeContext;
 pub use py_serialize::PyAnySerializer;