+++ crates/ryo3-serde/src/ser/py_types/py_bool.rs
@@ -1,9 +1,7 @@
 use pyo3::prelude::*;
 use serde::ser::{Serialize, Serializer};
-use std::ptr;
 
-use crate::ser::traits::PySerializeUnsafe;
-use pyo3::{Borrowed, ffi};
+use pyo3::Borrowed;
 
 pub(crate) struct PyBoolSerializer<'a, 'py> {
     obj: Borrowed<'a, 'py, PyAny>,
@@ -25,7 +23,7 @@ impl Serialize for PyBoolSerializer<'_, '_> {
     {
         #[expect(unsafe_code)]
         unsafe {
-            let istrue = ptr::eq(self.obj.as_ptr(), ffi::Py_True());
+            let istrue = std::ptr::eq(self.obj.as_ptr(), pyo3::ffi::Py_True());
             serializer.serialize_bool(istrue)
         }
     }
@@ -44,18 +42,5 @@ impl Serialize for PyBoolSerializer<'_, '_> {
             .map_err(crate::errors::pyerr2sererr)?
             .is_true();
         serializer.serialize_bool(tf)
-    }
-}
-
-impl PySerializeUnsafe for PyBoolSerializer<'_, '_> {
-    fn serialize_unsafe<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        #[expect(unsafe_code)]
-        unsafe {
-            let istrue = ptr::eq(self.obj.as_ptr(), ffi::Py_True());
-            serializer.serialize_bool(istrue)
-        }
     }
 }