+++ crates/ryo3-serde/src/ser/py_types/py_float.rs
@@ -1,10 +1,6 @@
+use pyo3::Borrowed;
 use pyo3::prelude::*;
 use serde::ser::{Serialize, Serializer};
-
-use crate::errors::pyerr2sererr;
-
-use crate::ser::traits::PySerializeUnsafe;
-use pyo3::Borrowed;
 
 pub(crate) struct PyFloatSerializer<'a, 'py> {
     obj: Borrowed<'a, 'py, PyAny>,
@@ -16,27 +12,29 @@ impl<'a, 'py> PyFloatSerializer<'a, 'py> {
     }
 }
 
+#[cfg(not(any(PyPy, GraalPy, Py_LIMITED_API)))]
 impl Serialize for PyFloatSerializer<'_, '_> {
     #[inline]
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
-        let v: f64 = self.obj.extract().map_err(pyerr2sererr)?;
-        serializer.serialize_f64(v)
+        // WENODIS: this is fo sho a float bc we cannot have gotten here w/o it being a float
+        #[expect(unsafe_code)]
+        let f = unsafe { self.obj.cast_unchecked::<pyo3::types::PyFloat>() }.value();
+        serializer.serialize_f64(f)
     }
 }
 
-impl PySerializeUnsafe for PyFloatSerializer<'_, '_> {
+#[cfg(any(PyPy, GraalPy, Py_LIMITED_API))]
+impl Serialize for PyFloatSerializer<'_, '_> {
     #[inline]
-    fn serialize_unsafe<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
     {
-        use pyo3::ffi::PyFloat_AsDouble;
-
-        #[expect(unsafe_code)]
-        let f = unsafe { PyFloat_AsDouble(self.obj.as_ptr()) };
-        serializer.serialize_f64(f)
+        use crate::errors::pyerr2sererr;
+        let v: f64 = self.obj.extract().map_err(pyerr2sererr)?;
+        serializer.serialize_f64(v)
     }
 }